+++ tests/Feature/Auth/AuthenticationTest.php
@@ -64,7 +64,9 @@ class AuthenticationTest extends TestCase
 
         $response
             ->assertOk()
-            ->assertSeeVolt('layout.navigation');
+            ->assertSee('Courses')
+            ->assertSee('My Courses')
+            ->assertSee('Profile');
     }
 
     public function test_users_can_logout(): void